requirements.txt
torch
transformers
datasets
numpy
scikit-learn
tqdm
pyyaml
pandas
